+++ src/MarkdownTable.ts
@@ -14,23 +14,26 @@ export class MarkdownTable {
 	readonly header!: string[];
 	readonly ranges: Map<number, Range[]> = new Map();
 	format: string[] = [];
-	readonly body: string[][] = [];
+	readonly body?: string[][] = [];
 	readonly defaultBody: string[][] = [];
 	readonly range: Range;
 	readonly headerRange: Range;
+	readonly formatRange: Range;
 	private _columnSizes: number[] = [];
 	private rawContainPipes = false
+	private _isInList = false
+	private _listIndentation: string[] = []
 
 	get id(): string {
 		return this._id;
 	}
 
 	get columns(): number {
-		return this.body[0].length;
+		return this.header.length;
 	}
 
 	get bodyLines(): number {
-		return this.body.length;
+		return this.body?.length || 0;
 	}
 
 	get totalLines(): number {
@@ -48,6 +51,14 @@ export class MarkdownTable {
 		return this.start.line;
 	}
 
+	get isInList(): boolean {
+		return this._isInList;
+	}
+
+	get listIndentation(): string[] {
+		return this._listIndentation;
+	}
+
 	constructor(start: Position, end: Position, regexpArray: XRegExp.ExecArray) {
 		this.start = start;
 		this.end = end;
@@ -56,9 +67,16 @@ export class MarkdownTable {
 
 		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 		const groups = regexpArray.groups!
-		this.rawContainPipes = [groups.header, groups.format, ...groups.body.split('\n').filter(x=>{return x.trim().length > 0})].every(v => { 
-			return v.search("^\\s*\\|") >= 0 && v.search("\\|\\s*\\r?\\n?$") > 0;
-		});
+
+		if (groups.body) {
+			this.rawContainPipes = [groups.header, groups.format, ...groups.body?.split('\n').filter(x => { return x.trim().length > 0 })].every(v => {
+				return v.search("^\\s*\\|") >= 0 && v.search("\\|\\s*\\r?\\n?$") > 0;
+			});
+		} else {
+			this.rawContainPipes = [groups.header, groups.format].every(v => {
+				return v.search("^\\s*\\|") >= 0 && v.search("\\|\\s*\\r?\\n?$") > 0;
+			});
+		}
 
 		this.headerRange = new Range(this.start, new Position(this.start.line, groups.header.length));
 
@@ -80,35 +98,46 @@ export class MarkdownTable {
 
 		this.format = splitCells(stripHeaderTailPipes(groups.format));
 		this.format = this.format.map(x => { return x.replace('\n', '') });
-		// this.format = this.format.filter(x => { return x.trim().length > 0 })
+		this.formatRange = new Range(this.start.with(this.start.line + 1), new Position(this.start.line + 1, groups.format.length));
 
 		firstLine += 1;
 
-		this.body = groups.body.replace(/^\r?\n+|\r?\n+$/g, '').split(/\r?\n/).map((lineBody: string) => {
-			let result = splitCells(stripHeaderTailPipes(lineBody));
-			result = result.map(x => { return x.replace('\n', '') });
-			// result = result.filter(x => { return x.trim().length > 0 })
-			return result
-		});
-		this.body.forEach((line: string[], line_index) => {
-			line.forEach((h, index) => {
-				const length = h.length;
-				const start = groups.body.split(/\r?\n/)[line_index].indexOf(h);
-				if (!this.ranges.has(index)) {
-					this.ranges.set(index, []);
-				}
-				this.ranges.get(index)?.push(new Range(new Position(firstLine, start), new Position(firstLine, start + length)));
+		if (groups.body) {
+			this.body = groups.body.replace(/^\r?\n+|\r?\n+$/g, '').split(/\r?\n/).map((lineBody: string) => {
+				let result = splitCells(stripHeaderTailPipes(lineBody));
+				result = result.map(x => { return x.replace('\n', '') });
+				return result
 			});
-			firstLine += 1;
-		});
-
-		if (groups.header) {
+			this.body.forEach((line: string[], line_index) => {
+				line.forEach((h, index) => {
+					const length = h.length;
+					const start = groups.body.split(/\r?\n/)[line_index].indexOf(h);
+					if (!this.ranges.has(index)) {
+						this.ranges.set(index, []);
+					}
+					this.ranges.get(index)?.push(new Range(new Position(firstLine, start), new Position(firstLine, start + length)));
+				});
+				firstLine += 1;
+			});
 			this._columnSizes = columnSizes(this.header, this.body);
 		} else {
-			this._columnSizes = columnSizes(this.body[0], this.body);
+			this._columnSizes = columnSizes(this.header);
 		}
 
 		this._id = md5(this.startLine.toString());
+
+		if (groups.inlist !== undefined) {
+			this._isInList = true
+			this._listIndentation[0] = groups.inlist + " "
+			for (let index = 1; index < this.totalLines; index++) {
+				this._listIndentation.push(" ".repeat(groups.inlist.length + 1))
+			}
+		} else if (groups.indentation !== undefined) {
+			this._isInList = true
+			for (let index = 0; index < this.totalLines; index++) {
+				this._listIndentation.push(groups.indentation)
+			}
+		}
 	}
 
 	sortedByColumn = (): MarkdownTableSortOptions => {
@@ -130,9 +159,9 @@ export class MarkdownTable {
 
 	notFormatted = (): string => {
 		const addTailPipesIfNeeded = this.rawContainPipes ? addTailPipes : (x: string) => x;
-		let joined = [this.format, ...this.body].map(joinCells).map(addTailPipesIfNeeded);
+		let joined = [this.format, ...this.body!].map(joinCells).map(addTailPipesIfNeeded);
 		if (this.header) {
-			joined = [this.header, this.format, ...this.body].map(joinCells).map(addTailPipesIfNeeded);
+			joined = [this.header, this.format, ...this.body!].map(joinCells).map(addTailPipesIfNeeded);
 		}
 		return joined.join('\n');
 	};