+++ src/MarkdownTableRegex.ts
@@ -1,26 +1,26 @@
 import XRegExp = require('xregexp');
 XRegExp.install("namespacing");
+
 export const tableRegex = XRegExp(String.raw`
-(?<header>                                # header capture
-  ^                                       # line start
-  (?:[^\r\n]*?\|[^\r\n]*)                 # line w/ at least one pipe
-  $                                       # line end
-  (?:\r?\n)                               # newline
-)?                                        # optional header
-(?<format>                                # format capture
-  ^                                       # line start
-  (?:[ :-]*?\|[ :-]*)+                    # line containing separator items w/ at least one pipe
-  \s*                                     # maybe trailing whitespace
-  $                                       # line end
-  (?:\r?\n)?                              # newline
+(?:
+  (?<inlist>[-+*]|[0-9]+\.)                  # Table is inside list
+  |                                          # or
+  (?<indentation>[ ]+)                       # Table is indented
+)?
+(?<header>                                   # Header capture
+  (?:[^\r\n]*?\|[^\r\n]*)+                   # Line w/ at least one pipe
+  \r?\n                                      # Newline
 )
-(?<body>                                  # body capture
-  (?:
-    (?:[^\r\n]*?\|[^\r\n]*)               # line w/ at least one pipe
-    \s*?                                  # maybe trailing whitespace
-    (?:\r?\n|^|$)?                        # newline
-  )+                                      # at least one
+(?<format>                                   # Format capture
+  (?:[ :-]*?\|[ :-]*)+                       # Line w/ separators containing at least one pipe
+  (?:\r?\n|$)                              # Newline
 )
+(?<body>
+  (?:                                        # Body capture
+    (?:[^\r\n]*?\|[^\r\n]*)+                 # Line w/ at least one pipe
+    (?:\r?\n|$)                              # Newline
+  )+
+)*
 `,
-	'gmx',
+    'gmx',
 );